# apps/api/src/app/schemas/browser_info.py
"""ブラウザ情報ツール向けのレスポンススキーマ。"""

from pydantic import BaseModel


class BrowserInfoResponse(BaseModel):
    """サーバーから見たクライアント情報。"""

    client_ip: str | None