# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ツールサービス API（バックエンド）",
    "version": "0.1.0"
  },
  "paths": {
    "/auth/sso/verify": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Sso Verify",
        "description": "SSO が発行したコードを検証し、結果を返す。",
        "operationId": "sso_verify_auth_sso_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SsoVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SsoVerifyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Auth Me",
        "description": "アクセストークンから取得した、現在ログイン中のユーザー情報を返す。",
        "operationId": "auth_me_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/tools/browser-info": {
      "get": {
        "tags": [
          "tools"
        ],
        "summary": "Get Browser Info",
        "description": "サーバーから見たクライアント IP を返す。",
        "operationId": "get_browser_info_tools_browser_info_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrowserInfoResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/tools/download-speed/chunk": {
      "get": {
        "tags": [
          "tools"
        ],
        "summary": "Get Download Chunk",
        "description": "指定サイズのランダムバイト列を octet-stream で返す。",
        "operationId": "get_download_chunk_tools_download_speed_chunk_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "size_mb",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Size Mb"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tools/summarize": {
      "post": {
        "tags": [
          "tools"
        ],
        "summary": "Post Summarize",
        "description": "Google Gemini で入力テキストの要約を返す。",
        "operationId": "post_summarize_tools_summarize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SummarizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummarizeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/health": {
      "get": {
        "summary": "Health Check",
        "description": "API の生存確認に用いる簡易エンドポイント。",
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Health Check Health Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AuthenticatedUser": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roles"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "roles"
        ],
        "title": "AuthenticatedUser",
        "description": "ログイン済みユーザーとして公開する情報。"
      },
      "BrowserInfoResponse": {
        "properties": {
          "client_ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Ip"
          }
        },
        "type": "object",
        "required": [
          "client_ip"
        ],
        "title": "BrowserInfoResponse",
        "description": "サーバーから見たクライアント情報。"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MeResponse": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthenticatedUser"
          }
        },
        "type": "object",
        "required": [
          "user"
        ],
        "title": "MeResponse",
        "description": "`/auth/me` のレスポンス。"
      },
      "SsoVerifyRequest": {
        "properties": {
          "code": {
            "type": "string",
            "minLength": 1,
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "SsoVerifyRequest",
        "description": "`/auth/sso/verify` に送るリクエストボディ。"
      },
      "SsoVerifyResponse": {
        "properties": {
          "accessToken": {
            "type": "string",
            "title": "Accesstoken"
          },
          "user": {
            "$ref": "#/components/schemas/AuthenticatedUser"
          }
        },
        "type": "object",
        "required": [
          "accessToken",
          "user"
        ],
        "title": "SsoVerifyResponse",
        "description": "`/auth/sso/verify` のレスポンス。"
      },
      "SummarizeRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 10000,
            "minLength": 1,
            "title": "Text"
          },
          "length": {
            "type": "string",
            "enum": [
              "short",
              "medium",
              "long"
            ],
            "title": "Length",
            "default": "medium"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "SummarizeRequest",
        "description": "`POST /tools/summarize` のリクエストボディ。"
      },
      "SummarizeResponse": {
        "properties": {
          "summary": {
            "type": "string",
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "summary"
        ],
        "title": "SummarizeResponse",
        "description": "`POST /tools/summarize` のレスポンスボディ。"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}